+++ src/Legacy/Message/RequestContract.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace Rougin\Datatables\Legacy\Message;
+
+interface RequestContract
+{
+    public function columns();
+
+    public function draw();
+
+    public function length();
+
+    public function orders();
+
+    public function search();
+
+    public function start();
+}