+++ src/Legacy/Request/ColumnContract.php
@@ -1,0 +1,16 @@
+<?php
+
+namespace Rougin\Datatables\Legacy\Request;
+
+interface ColumnContract
+{
+    public function data();
+
+    public function name();
+
+    public function orderable();
+
+    public function search();
+
+    public function searchable();
+}