+++ src/Legacy/Request/OrderContract.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rougin\Datatables\Legacy\Request;
+
+interface OrderContract
+{
+    public function ascending();
+
+    public function column();
+
+    public function descending();
+}