+++ src/Legacy/Request/SearchContract.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Rougin\Datatables\Legacy\Request;
+
+interface SearchContract
+{
+    public function regex();
+
+    public function value();
+}