+++ src/Legacy/Request/SearchCreation.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Rougin\Datatables\Legacy\Request;
+
+interface SearchCreation
+{
+    public function make();
+
+    public function regex($regex);
+
+    public function value($value);
+}