+++ CHANGELOG.md
@@ -5,4 +5,4 @@ All notable changes to `Datatables` will be documented in this file.
 ## 0.1.0 - Unreleased
 
 ### Added
-- `Datatables` library
+- `Datatables` package